backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
pydantic==2.5.3
python-dateutil==2.8.2
aiosqlite==0.19.0
